Move new AssetHandle class to a different file
