+++ earth_enterprise/src/fusion/autoingest/StorageManagerAssetHandle.h
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2019 The Open GEE Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef STORAGEMANAGERASSETHANDLE_CPP
+#define STORAGEMANAGERASSETHANDLE_CPP
+
+#include <functional>
+#include <memory>
+
+#include "autoingest/.idl/storage/AssetDefs.h"
+
+template<class AssetType>
+using AssetPointerType = std::shared_ptr<AssetType>;
+
+// Objects outside the storage manager will access assets through AssetHandles.
+// This allows the storage manager to properly clean up when the other object
+// is done using the asset (release locks, update cache size, etc).
+//
+// This AssetHandle will eventually replace the AssetHandle_ definied in
+// AssetHandle.h and its derivatives.
+template<class AssetType>
+class AssetHandle {
+  private:
+    template<class OtherAssetType> friend class AssetHandle;
+    AssetPointerType<AssetType> ptr;
+    std::function<void(void)> onFinalize;
+  public:
+    AssetHandle(AssetPointerType<AssetType> ptr,
+                std::function<void(void)> onFinalize)
+        : ptr(ptr), onFinalize(onFinalize) {}
+    AssetHandle() = default;
+    // Assignment from handles with convertable pointers
+    template<class OtherAssetType>
+    AssetHandle(const OtherAssetType &other)
+        : ptr(std::dynamic_pointer_cast<AssetType>(other.ptr)), onFinalize(other.onFinalize) {
+      assert(ptr);
+    }
+    ~AssetHandle() {
+      if (onFinalize) onFinalize();
+    }
+    inline AssetType * operator->() const { return ptr.operator->(); }
+    inline explicit operator bool() const { return ptr && (ptr->type != AssetDefs::Invalid); }
+};
+
+#endif // STORAGEMANAGERASSETHANDLE_CPP
